extraction history hooked with backend
